SLProgram: add a .ptr field, to keep track of next store location
We want to support GAP's overwrite feature: the assign op will
allow to overwrite a previous result, which means that we don't
always increase the size of `res` during evaluation, which means
that the current length of `.lines` during SLProgram building
is _not_ necessarily matching the corresponding position in `res`
which will be written. `.ptr` indicates the next write location.

This will also support "deleting" (trailing) positions in `res`,
to erase intermediate calculations and make it easier to
convert from GAP's SLPs.