add Free SLPs (and demote Lazy as an SLP)
One reason is that when `Lazy <: AbstractSLProgram`,
then concrete subtype of `Lazy` are also
`<: AbstractSLProgram`, which doesn't work well.
Also, combining two Lazy's can change the meaning
of the `Gen`s (i.e. evaluate doesn't know the
index corresponding to a specific `Gen` unless
a list of symbols is passed.