automorphisms of QabField()
implementation of automorphisms defined by \zeta_n \mapsto \zeta_n^k;
for example, `QabAutomorphism(-1)` is complex conjugation.

(Creating the coefficients list and then the field element
is better than doing arithmetics with field elements or polynomials.
It should be even cheaper to create the element in the cyclotomic field `F`
directly via `F(coeffs)` instead of having an intermediate step `R(coeffs)`
where `R` is the underlying polynomial ring, but this seems to be not allowed.)