Radical of ideals
I added a function to compute the radical of an ideal, including a test.
On the way I added a constructor of type "ideal(MPolyRing, sideal)", since this one was missing
and one of type "(MPolyRing)(spoly)" in order to be able to construct Oscar ideals from arrays of spolys.