Further functionality for AlgHoms
Added a test for finiteness, added the computation of an inverse map, if existent,
and added the computation of a preimage of an element, if it is in the image.
The preimage function returns, for the moment, a tuple containing the element in
the preimage and the Kernel.
Changed the structure of AlgHoms by storing the groebner data involved in the computations
regarding surjectivity, preimages, inversion and finiteness. Addapted existing code.