Transform form groups (#368)
In this PR I added methods to relate forms and groups of isometries preserved by the forms. The relevant novelties are in the file `src/Groups/matrices/forms.jl`, in the last two bunch of methods.
In the first bunch, there are functions to get the forms preserved by a group of matrices.
In the second bunch, the functions to compute the group of isometries of a given sesquilinear or quadratic form.

About the first one:
three approaches are shown. Which is the best one? We can also decide to keep more than one (e.g. sometimes one needs just any form preserved by the group, sometimes he may need all the forms).
Note that the first approach finds forms with every possible matrix, not just symmetric / hermitian / skew-symmetric, but I want to write a function to restrict to one of these cases.

About the second one:
this works only with sesquilinear or quadratic forms, so forms with Gram matrix being symmetric / hermitian / skew-symmetric. It works also with degenerate matrices.