Change Array -> Vector / Matrix
In Julia 1.5 and later, the type of 1 resp. 2 dimensional arrays is printed as
Vector resp. Matrix, and thus doctests need to be adjusted accordingly. The
easiest way to do that was to change this with a regex everywhere, which also
modifies code. I think this makes sense for consistency reasons, though, and
also seems in line with the vision of the Julia team.