Primality proving (#470)
* Functions for elliptic curves
Added functions for elliptic curves based on the corresponding functions in Hecke

* Elliptic Curve Method
Added a structure for elliptic curves over Z/nZ (given in short Weierstrass form), added a function for the Elliptic Curve Method and points addition for elliptic curves over Z/nZ

* Changed the input of ECM from Int64 to fmpz

* Updated the corresponding test

* Improved multiplication with constant by using repeated squaring.

* Removed unused functions from export

* Primality Proving
Added functions for Miller-Rabin test, Pollard rho, Pollard p-1 and functions related to the Elliptic Curve Primality Proving. The ECPP test seems to run for numbers up to 40 digits. It is not competitive, but it could be used for teaching purposes.