docs build: use temporary environment for docs project by default
copy existing Project and develop all required packages in Oscar.doc_init
to make sure everything is loaded from the same paths as in the currently
loaded Oscar packages.
also avoids some resolve-issues and issues arising from outdated manifest files in docs/

There is also a flag to specify a custom directory for the docs-project which is used by the CI, but will modify the docs/Project.toml file.