get rid of the `mat_iso` field in matrix groups
@fieker had observed that an individual `mat_iso` field is not needed in matrix groups and forms.
What we need is the possibility to translate matrices from Oscar to GAP and back,
and this can be done generically via the known `ring_iso` that translates matrix entries.
The recommended user function for the direction from Oscar to GAP is `map_entries`.
(We cannot use the available `map_entries` method because `parent`and `_change_base_ring` do not fit to the situation in GAP.)
For the direction from GAP to Oscar, we introduce `preimage_matrix`:
The former calls of the form `preimage(G.mat_iso, x)` got replaced by `preimage_matrix(G.ring_iso, x)`.

In fact, the direction from GAP to Oscar seems to be needed only in few places.
When one constructs a matrix group element from a GAP matrix,
one can create the Oscar matrix (field `elm`) only on demand.
This feature had not really been used up to now.
With the current changes, the computations via `preimage_matrix` get delayed until someone accesses the `elm` field.