CI: don't kill PR CI runs for our own branches anymore
Now that we have the concurrency feature, these additional builds are not as
problematic as they used to be anymore: there can be at most one build per PR
and one per branch (except for the master branch). So there will be at most
two parallel builds for each such branch with an active PR.

On the other hand, disabling the PR CI builds was problematic, because they
actually differ from the branch CI builds: the former are performed with a
temporary merge commit of the branch into master; the latter with just the tip
of the branch. But we really want and need the former behavior to be sure
a merge won't break master.