construct subgroups compatible with the given group
(This addresses one item from #894.)

Calls of the form `T(gapobj)`, for a type `T`, are not suitable
when the result shall be compatible with a given group.
In order to set the correct degree (in case of a permutation group)
or the correct base ring information (in case of a matrix group),
one can call `_as_subgroup_bare`.