fix conjugation of perm. groups by permutations
This fixes the second problem mentioned in issue #894,
that is, conjugating a group by an element preserves the degree.

(The fix works only for permutation groups,
I do not see how to formulate a generic method.
Eventually we have to define what compatibility means
for several groups and group elements,
perhaps via a function `iscompatible` that returns a group
w.r.t. which one can then use as the first argument of `_as_subgroup_bare`?)