Adjust doctest for `describe(G::GAPGroup)`
The output of `describe` is in general not unique and depends on the
state of a random number generator. The previous doctest used an
example that was affected by this, which we remove.

Also add an example involving a free group, just to show this other
flavor of the function.