support `:indicators` and `:OD` in char. table display
Specifying `:indicators` in the `IOContext` of `show` for
character tables forces additional columns of row labels
that contain the indicator values in question.

Analogously, specifying `:OD` forces one additional column of row labels
that contain the orthogonal discriminants,
provided that the database of these values is available;
in general the values cannot be computed from the given character table.