admit result `GrpAbFinGen` for `maximal_abelian_quotient`
We will have the idea to cache the results of `maximal_abelian_quotient`,
parametrized by their types.
Then the question arises how to deal with the unary version,
where up to now the result of the GAP function determines the type of the
result in Oscar.
One natural solution would be to take any cached result if there is one,
which would make the description of the result type even more complicated.