Better type stability for MPolyQuo
... by removing the member `R` from the MPolyQuo -- we can instead use
the `base_ring` of the ideal `I`, which is type stable, plus it saves us
a bit of storage (not that it really matters here).

Also change code to use `base_ring(A)` instead of `A.R`, and update some
code to use accessor functions for MPolyQuo instead of going through
`base_ring`.