makeing progress, but still some bug
however, this is now definitely back to the grometry group as I cannot
see if the test, the implementation or the complexes are at fault.
You need to use the Complexes branch/ pr in Hecke as the complexes still
had some indexing problems.
Then the test run up to homology, but ext fails:

 Got exception outside of a @test
  AssertionError: !(isdefined(M, :quo)) && !(isdefined(U, :quo))
  Stacktrace:
    [1] quo(M::SubQuo{fmpq_mpoly}, U::SubQuo{fmpq_mpoly}, task::Symbol)
      @ Oscar ~/.julia/dev/Oscar/src/Modules/UngradedModules.jl:3012
    [2] homology(C::ChainComplex{ModuleFP}, i::Int64)
      @ Oscar ~/.julia/dev/Oscar/src/Modules/UngradedModules.jl:6028
    [3] ext(M::SubQuo{fmpq_mpoly}, N::SubQuo{fmpq_mpoly}, i::Int64)
      @ Oscar ~/.julia/dev/Oscar/src/Modules/UngradedModules.jl:6090