Replace '@doc Markdown.doc' by '@doc raw'
As far as I can tell, there is no functional difference.

We could now also remove all remaining references to 'Markdown', but that
might break various in-flight PRs. So I only removed some, and also kept
Markdown in Project.toml. We can fully remove it in a future PR.

Most of the changes in this PR were made with this command:

    git grep -l Markdown | xargs perl -pi -e 's;\@doc Markdown.doc""";\@doc raw""";g'