make sure that certain GAP packages are loaded
Up to now, `GAP.Packages.load` is called for several GAP packages
in Oscar's `__init__`, but we do not check whether the GAP package
is really loaded afterwards.
(This means "load the package if it is available, but do not complain if it isn't".)

Now we throw an error if the packages cannot be loaded.
Hence the Oscar code can assume that they are available, and need not check for it.

The GAP packages in question do not require compilation, thus I do not expect complications at runtime because of this change.
(This is the reason why the packages `Browse` and `ferret` are currently not forced to be loaded.)