[PolyhedralGeometry] star_subdivision improvement
It is already clear that the implementation assembles a ray that is not
a positive scalar multiple of an existing ray, as it lives inside a cone
tau of the fan.

So the question is whether the new cones constructed are actually
maximal. This is also true, since we assume all cones containing our
newly subdivided cone tau to be smooth. For every maximal cone M
containing tau we subtract all rays of tau and then add a facet of tau
with the new ray.  This will also be maximal, since the new ray is truly
new as described in the beginning and tau and M were smooth.

Closing #2286