Run the doctests in fully seperate jobs
We loose running doctests on macOS this way, but I think that's an
acceptable compromise: it's unlikely the results for those will differ
on macOS; indeed we mostly test on macOS to verify JLLs work there,
and to catch OS specific regressions on nightly.

If we truly want it, we could re-enable it with some effort, but for
now it doesn't seem worth it to me.