changed the setup for `iso_gap_oscar`
Let GAP's method selection decide which Julia function does the work.
This way, one can install methods for other GAP domains
than the ones that are treated in `src/GAP/iso_gap_oscar.jl`.

If the method gets defined before the OscarInterface gets loaded,
it is sufficient to add it to the global array `_iso_gap_oscar_methods`,
together with the description of the GAP filter to which the method belongs;
this applies to code in `Oscar.jl/experimental`.
(Otherwise one has to call GAP's `InstallMethod`.)