CI: avoid timeouts on macOS
To conserve resources, we were running both the regular test suites and
the doctest on a single macOS instance. That was good until the test
suite became so slow that this started routinely time out. That means
that now those macOS jobs run ~150 minutes just to fail.

That's a waste, as those macOS runner are strictly limited and often
pose the bottleneck for our CI tests to run.

So until the test suite gets faster again (if ever), revert to just
running the doctests. Together with the regular tests being run under
macOS with Julia nightly, plus several of us developing on macOS, this
should hopefully still catch the worst and most obvious macOS
regressions.

If this turns out problematic, we'll have to consider alternatives, including:
1. using a third macOS runner again,
2. increasing the timeout
3. using a custom GitHub runner, using a macOS machine we own and control,