Revise some character table tests to use a doctest (#2824)
These tests deal with testing how character tables are printed
in various situations. We change this to (ab)use the `jldoctest`
system from Documenter. This brings with it the advantage that it
is very easy to adjust the tests for desired changes to the printing:
it can be done changing

    doctest(nothing, [AuxDocTest_GroupCharacters])

to

    doctest(nothing, [AuxDocTest_GroupCharacters]; fix=true)

and then running this in the Julia REPL:

    include("test/Groups/group_characters.jl")