support `gen(G::GAPGroup, 0)`
This was already supported for `FinGenAbGroup`.

(The delegations between methods are not the same:
For `FinGenAbGroup`, there is special code for `getindex(A, i)`,
and `gen(A, i)` calls `getindes(A, i)`.
For `GAPGroup`, there is special code for `gen(G, i)`,
and the generic `getindex` method from AbstractAlgebra calls `gen(G, i)`.)