Tweak `exterior_algebra` tests
Do not just verify that certain `exterior_algebra` invocations don't error
out; also check that the returned values satisfy certain properties.

Also disable two `@test_broken` tests for now, because (a) the underlying
issue is fixed in a Singular PR, but this causes OscarCI failures there.. (b)
even if the underlying issue is fixed, those test were wrong, because
`exterior_algebra` doesn't return a boolean, but the return value was used in
a context where it must be boolean (namely as "argument" for `@test_broken`).