Rename Spec to AffineScheme #3345 (#3425)
* Rename Spec -> AffineScheme

Use `AffineScheme` only for the type, not the function.

* Rename proj for toric divisors to projectivization

As was recommended by Martin Bies and Wolfram Decker.

* Function ProjectiveScheme to projective_scheme and proj

* Fix: spec now extends the constructor AffineScheme

In particular, spec allows the argument to be a quotient ring or localized ring. Similarly for proj. Probably the tests would not have passed in the previous commit because of this

* add support for projective_scheme and affine_scheme

---------

Co-authored-by: Simon Brandhorst <brandhorst@math.uni-sb.de>
(cherry picked from commit b00d5e444e1c353ea833340f190fffc522ebabe1)