Add GapObj / julia_to_gap conversion methods for more types
Specifically, for
- MatrixGroup
- MatrixGroupElem
- SesquilinearForm
- GroupCoset
- GroupDoubleCoset
- GAPGroupConjClass

I left out SubgroupTransversal because there are additional
semantics on the Julia side invisible to the GAP object, so the
"conversion" is not always quite safe.

Also did nothing for GAPGroupClassFunction as there is no obvious
"free" conversion available there. We could still add a method,
but I'll leave that to a potential future PR.