fix relation between `PcGroup` and `pc_group`
`PcGroup` takes a GAP group and wraps it;
it is not allowed to replace the group by a different GAP group,
thus it must throw an exception if the GAP group is not suitable.
(This is crucial for example in the creation of group homomorphisms,
where the GAP objects stored in domain and codomain must fit to the
`Source` and `Range` objects of the underlying GAP mapping.)

`pc_group` may return an Oscar group that stores a different
GAP group than the given one.