change `quo` to prefer a `PcGroup` result to a `SubPcGroup`
I am not sure whether this is a good idea.

On the one hand, users would not understand that a newly created
quotient is represented as a `SubPcGroup`.
(The starting point was changed output in an example for the Oscar book.)

On the other hand, we have to work around a GAP feature that offers
natural homomorphisms instead of natural epimorphisms (with the
possibility to create maps that need not be surjective).

Since we have to compute the image of the GAP mapping anyhow,
the current change does not add expensive computations.