No type piracy: remove matrix*vector methods for Nemo types (#760)
* No type piracy: remove matrix*vector methods for Nemo types

* Adjut tests

* Remove more problematic * methods

Multiplying FreeModuleElem by a matrix group element returning a
Vector{} is weird and shouldn't be done. While this could be fixed
by adding a second transpose call, I don't think this kind of
inefficient operation should be encouraged, that vector space
simply is a right-module, end of story.

In a similar vein, remove the three argument * method.