make `isomorphism(PcGroup, A)` for infinite abelian `A` work
The documentation claimed that this is not supported,
but in fact GAP's pcp groups can deal with this case.

The code looks horrible,
due to the fact that GAP supports the decomposition
of elements in abelian groups w.r.t. independent generators
only for special such generating sets.
(This happens for abelian permutation groups.)

Note that this is generic code.
Perhaps we should add special methods for the case that we ask for
a `FPGroup` or a `PcGroup`, then also the functions for computing
(pre)images can be more efficient.