Remove toric blowups along ideals
Instead, have a function to check whether the blowup can be a toric blowup. This function is currently very specialized. It only works when all the generators of the ideal are generators of the Cox ring. Therefore, I moved this function to FTheoryTools, the only place where, I believe, this is currently used.

This also fixes a bug in the previous check. Namely, previously the blowup along an ideal could give wrong results. In the example below, the last blowup is along an ideal containing the irrelevant ideal. This is a blowup along the empty subscheme, so it should be an isomorphism, but it is not. In the new code, this is fixed.