fix the Julia functions used by GAP, improve the test
That was subtle:
By default, the Julia function used for embedding a matrix
into a matrix group makes a membership test.
We use the embedding in order to write the functions that map
characteristic zero matrices in GAP to matrices over finite fields
in GAP and vice versa.
These functions describe the `NiceMonomorphism` of the
characteristic zero matrix group in GAP,
hence they implement in particular the membership test on the GAP side.
Thus we have to omit the membership tests on the Julia side
in order to avoid infinite recursions.