Tweak printing of abelian_closure(QQ)
Instead of "Abelian closure of Q" now print "Abelian closure of Q"
resp. (in terse mode) "Abelian closure of QQ".

The old printing was a bit jarring in that it differed from how
algebraic closures are printed, and also the use of "Q" to
represent the rationals is not used anywhere (?) else in OSCAR.