Remove broken tropical Groebner basis shortcut for binomial ideals (#4697)
This broken shortcut was originally added in a joint project with
Henriksson and Helminck.  The idea was that groebner_basis should not
do anything for binomial ideals if the weight vector lies on the
tropicalization (the weight vector test was missing).  While this is
mathematically correct, it is on second thought not a good idea if
groebner_basis for binomial ideals behaves differently than for
general ideals, so I am taking this out.