improve the handling of matrix groups over number fields
(These changes require the GAP changes from gap-system/gap/pull/5970.)

The idea is to set the GAP filter `MayBeHandledByNiceMonomorphism`
in the `GapObj` of an Oscar matrix group over a number field,
and to store the corresponding Oscar group in the attribute `JuliaData`
of the GAP group, in order to give GAP access to the Oscar group.

As soon as the Oscar group asks its `GapObj` for a computation
that may benefit from a "nice monomorphism",
the GAP group is asked for its `IsHandledByNiceMonomorphism` value.
This question triggers the `is_finite` computation for the Oscar group.
If the answer is `false` then "nice monomorphisms" do not help,
and if the answer is `true` then an isomorphism from the Oscar group
to a matrix group over a finite field has been computed,
which is used to create a "nice monomorphism"  for the GAP group.

For that, a `IsHandledByNiceMonomorphism` method for the GAP group
is needed, which makes an `is_finite` test.
Note that up to now, `is_finite` for matrix groups simply delegated
the work to GAP,
which eventually computes a "nice monomorphism" in a different way,
and then the reduction provided by Oscar does not help.
Now we use a special `is_finite` method that triggers the computation
of the reduction to a matrix group over a finite field.
Since the reduction functions throw an error exception if (and only if)
the matrix group in question is not finite, we use `try`/`catch`.
Perhaps this setup should better be changed, in the sense that
the return value of `_isomorphic_group_over_finite_field` indicates
whether the given matrices generate a finite group or not.