Fix number theory booktest
The order of the generators of G changes with recent Julia versions,
so enforce that we pick the same group element across all versions
by selecting it based on one of its properties. Of course this still
relies on the order of the roots to stay fixed, but for now this
seems to be sufficient.