Fix spelling: this is OSCAR, not Oscar
We agreed on this as the canonical spelling and use it in the
majority of places, but new instances of the wrong spelling keep
creeping in, so we need to occassionally clean them up.

It is understandable that there is some confusion as the package
itself is called Oscar.jl. That's an unfortunate historical fact
that we can't easily change.

That said, text which explicitly refers to the Julia package or
module may of course still use `Oscar`.