Speed up _overlattice_stabilizer over fields
Before:

    julia> L = integer_lattice(gram=ZZ[2 1 -1 1 -1 -1 0 0 0 0 0 0 0 0 0 1; 1 2 -1 1 -1 0 0 0 0 0 0 0 0 0 0 1; -1 -1 2 0 1 0 0 0 0 0 0 0 0 0 0 -1; 1 1 0 2 -1 0 0 0 0 0 0 0 0 0 0 0; -1 -1 1 -1 2 0 0 0 0 0 0 0 0 0 0 0; -1 0 0 0 0 2 0 0 0 0 0 0 0 0 0 -1; 0 0 0 0 0 0 2 1 -1 1 -1 -1 -2 -1 -2 -1; 0 0 0 0 0 0 1 2 -1 1 -1 -1 -2 -1 -2 -1; 0 0 0 0 0 0 -1 -1 2 0 0 0 2 0 2 1; 0 0 0 0 0 0 1 1 0 2 -1 -1 -1 -1 -1 0; 0 0 0 0 0 0 -1 -1 0 -1 2 1 1 1 2 1; 0 0 0 0 0 0 -1 -1 0 -1 1 2 1 1 1 1; 0 0 0 0 0 0 -2 -2 2 -1 1 1 4 1 3 2; 0 0 0 0 0 0 -1 -1 0 -1 1 1 1 2 0 0; 0 0 0 0 0 0 -2 -2 2 -1 2 1 3 0 6 3; 1 1 -1 0 0 -1 -1 -1 1 0 1 1 2 0 3 4]);

    julia> G,_ = Oscar._isometry_group_via_decomposition(L);

    julia> S1,S2,_ = Hecke._shortest_vectors_sublattice(L);

    julia> Oscar.set_seed!(235) ; @time Oscar._overlattice_stabilizer(G, S1,S2);
      4.449402 seconds (5.96 M allocations: 204.601 MiB, 3.14% gc time)

After

    julia> L = integer_lattice(gram=ZZ[2 1 -1 1 -1 -1 0 0 0 0 0 0 0 0 0 1; 1 2 -1 1 -1 0 0 0 0 0 0 0 0 0 0 1; -1 -1 2 0 1 0 0 0 0 0 0 0 0 0 0 -1; 1 1 0 2 -1 0 0 0 0 0 0 0 0 0 0 0; -1 -1 1 -1 2 0 0 0 0 0 0 0 0 0 0 0; -1 0 0 0 0 2 0 0 0 0 0 0 0 0 0 -1; 0 0 0 0 0 0 2 1 -1 1 -1 -1 -2 -1 -2 -1; 0 0 0 0 0 0 1 2 -1 1 -1 -1 -2 -1 -2 -1; 0 0 0 0 0 0 -1 -1 2 0 0 0 2 0 2 1; 0 0 0 0 0 0 1 1 0 2 -1 -1 -1 -1 -1 0; 0 0 0 0 0 0 -1 -1 0 -1 2 1 1 1 2 1; 0 0 0 0 0 0 -1 -1 0 -1 1 2 1 1 1 1; 0 0 0 0 0 0 -2 -2 2 -1 1 1 4 1 3 2; 0 0 0 0 0 0 -1 -1 0 -1 1 1 1 2 0 0; 0 0 0 0 0 0 -2 -2 2 -1 2 1 3 0 6 3; 1 1 -1 0 0 -1 -1 -1 1 0 1 1 2 0 3 4]);

    julia> G,_ = Oscar._isometry_group_via_decomposition(L);

    julia> S1,S2,_ = Hecke._shortest_vectors_sublattice(L);

    julia> Oscar.set_seed!(235) ; @time Oscar._overlattice_stabilizer(G, S1,S2);
      0.444009 seconds (1.71 M allocations: 143.068 MiB, 16.66% gc time)

Multiple things are changed; and some of the changes may now be redundant:

- compute `BL` with less overhead (doesn't matter overall, though)
- compute stabilizer in a permutation group (via nice monomorphism)
- convert all matrices to GAP matrices, then perform the computation
  entirely on the GAP side (this avoids constant conversion between GAP
  and OSCAR matrices)

Another change is that instead of multiplying ZZMatrix with a
zzModMatrix or fpMatrix, it is currently better to first convert
the ZZMatrix then perform the multiplication. A proper fix for
this  already has been submitted as a Nemo PR.

But even that is suboptimal: one should not constantly convert the
same ZZMatrix into zzModMatrix; instead this should be done once
per generator, and then the result kept and reused. The GAP
functions for orbits, stabilizers etc. have an API for that: in
addition to specifying a list of generators `gens`, and an action
function `act`, one can also specify a second list `acts` matching
`gens`; then when the algorithms want to compute the action of,
say, `act(v, gens[i])` they instead compute `act(v, acts[i])`.
So in our context, `gens` would be a `Vector{ZZMatrix}` while
`acts` might be a `Vector{zzModMatrix}`.

I think we should extend the OSCAR API for orbit/stabilizer to
allow something equivalent. The result would still be very
slightly less efficient than driving everything "directly", but
it would be general enough to be used by someone without learning
GAP first.

Alas, I can't write such a method here right now because of an
annoying issue in the GAP library code. Basically, it requires in
one spot that the list `acts` consists of objects which are valid
inputs for the GAP function `Group`, which our matrices are not.
But I think we can fix that, but it'll require some more work.

For the non-prime case, not over a field, we really need this, as
GAP doesn't implement the Howell form.