Don't pass kwargs to exterior shift on concrete matrices
At the moment, the kwargs in exterior_shift_lv will be a row reduction
algorithm `rref` and a hypergraph `lower_uhg`. `rref` is not used when
working with concrete matrices in `exterior_shift`, and passing
`lower_uhg` is not understood. We now just don't pass any kwargs down to
this `exterior_shift` call.