CI: stop testing against Julia 1.11
This will allow us to free up ~135 GB on our action runner
in Kaiserslautern.

Also, it will reduce load on CI, as we'll run fewer tests.

Usually, if things work in 1.10 and 1.12 they should also work in 1.11.
There can be exception, but so be it. I'd expect people to either use
the latest Julia or the LTS version anyway.